@@ -5,7 +5,7 @@ class RateLimiter {
   RateLimiter(this.timeout);
 
   bool shouldFetch(String key) {
-    final DateTime lastFetched = timestamps[key];
+    final DateTime? lastFetched = timestamps[key];
     final DateTime now = DateTime.now();
 
     if (lastFetched == null) {
